+++ workloads/testpmd/common.sh
@@ -184,16 +184,21 @@ cleanup_network() {
 }
 
 deploy_operator() {
-  log "Removing benchmark-operator namespace, if it already exists"
-  oc delete namespace benchmark-operator --ignore-not-found
-  log "Cloning benchmark-operator from branch ${operator_branch} of ${operator_repo}"
-  rm -rf benchmark-operator
-  git clone --single-branch --branch ${operator_branch} ${operator_repo} --depth 1
-  (cd benchmark-operator && make deploy)
-  oc wait --for=condition=available "deployment/benchmark-controller-manager" -n benchmark-operator --timeout=300s
-  oc adm policy -n benchmark-operator add-scc-to-user privileged -z benchmark-operator
-  oc adm policy -n benchmark-operator add-scc-to-user privileged -z backpack-view
-  oc patch scc restricted --type=merge -p '{"allowHostNetwork": true}'
+  if [[ "${isBareMetal}" == "false" ]]; then
+    log "Removing benchmark-operator namespace, if it already exists"
+    oc delete namespace benchmark-operator --ignore-not-found
+    log "Cloning benchmark-operator from branch ${operator_branch} of ${operator_repo}"
+  else
+    log "Baremetal infrastructure: Keeping benchmark-operator namespace"
+    log "Cloning benchmark-operator from branch ${operator_branch} of ${operator_repo}"
+  fi
+    rm -rf benchmark-operator
+    git clone --single-branch --branch ${operator_branch} ${operator_repo} --depth 1
+    (cd benchmark-operator && make deploy)
+    oc wait --for=condition=available "deployment/benchmark-controller-manager" -n benchmark-operator --timeout=300s
+    oc adm policy -n benchmark-operator add-scc-to-user privileged -z benchmark-operator
+    oc adm policy -n benchmark-operator add-scc-to-user privileged -z backpack-view
+    oc patch scc restricted --type=merge -p '{"allowHostNetwork": true}'
 }
 
 deploy_workload() {
@@ -203,7 +208,6 @@ deploy_workload() {
   log "Deploying testpmd benchmark"
   envsubst < $CRD | oc apply -f -
   #envsubst < $CRD > /tmp/testpmd.yaml
-  exit
   log "Sleeping for 60 seconds"
   sleep 60
 }
@@ -312,4 +316,4 @@ deploy_perf_profile
 deploy_operator
 deploy_workload
 wait_for_benchmark
-cleanup_network
+#cleanup_network
